added rotate left and rotate right, etc.
I added an if statement to get the IR read to check if the mouse should rotate, move forward or in reverse when the IR sensor picks up a signal of a wall or not. I need to test rotateleft and rotateright. The ir values in the if statement for loop-function needs to be tested to see what the threshold is for mouse to rotate.